+++ frontend/src/components/User/UserAboutTab.tsx
@@ -23,17 +23,17 @@ const UserAboutTab = () => {
           setEditDetails(event.target.value);
         }}
       />
-      {editingDetails ? (
-        <Spinner className="w-4 h-4" />
-      ) : (
-        <button
-          type="button"
-          onClick={submitUserDetails}
-          className="cursor-pointer focus:outline-none text-white bg-green-700 hover:bg-green-800 focus:ring-4 focus:ring-green-300 font-medium rounded-lg text-sm px-5 py-2.5 mt-4"
-        >
-          Edit Info
-        </button>
-      )}
+      <button
+        type="button"
+        onClick={submitUserDetails}
+        className="cursor-pointer focus:outline-none text-white bg-green-700 hover:bg-green-800 focus:ring-4 focus:ring-green-300 font-medium rounded-lg text-sm px-5 py-2.5 mt-4"
+      >
+        {editingDetails ? (
+          <Spinner className="w-4 h-4" />
+        ) : (
+          <div>Edit Info</div>
+        )}
+      </button>
     </div>
   );
 };
